feat(portal): unbundle patient portal dashboard with React Suspense widgets and premium loading skeletons
